+++ contracts/axelar-gas-service/src/contract.rs
@@ -4,10 +4,8 @@ use crate::error::ContractError;
 use crate::event;
 use crate::interface::AxelarGasServiceInterface;
 use crate::storage_types::DataKey;
-use axelar_soroban_std::contract_traits::{
-    MigratableInterface, OwnableInterface, UpgradableInterface,
-};
-use axelar_soroban_std::{contract_traits, ensure, types::Token};
+use axelar_soroban_std::interfaces::{MigratableInterface, OwnableInterface, UpgradableInterface};
+use axelar_soroban_std::{ensure, interfaces, types::Token};
 
 #[contract]
 pub struct AxelarGasService;
@@ -16,7 +14,7 @@ pub struct AxelarGasService;
 impl AxelarGasService {
     /// Initialize the gas service contract with a gas_collector address.
     pub fn __constructor(env: Env, owner: Address, gas_collector: Address) {
-        contract_traits::set_owner(&env, &owner);
+        interfaces::set_owner(&env, &owner);
         env.storage()
             .instance()
             .set(&DataKey::GasCollector, &gas_collector);
@@ -34,7 +32,7 @@ impl MigratableInterface for AxelarGasService {
     type Error = ContractError;
 
     fn migrate(env: &Env, migration_data: ()) -> Result<(), ContractError> {
-        contract_traits::migrate::<Self>(env, || Self::run_migration(env, migration_data))
+        interfaces::migrate::<Self>(env, || Self::run_migration(env, migration_data))
             .map_err(|_| ContractError::MigrationNotAllowed)
     }
 }
@@ -46,7 +44,7 @@ impl UpgradableInterface for AxelarGasService {
     }
 
     fn upgrade(env: &Env, new_wasm_hash: BytesN<32>) {
-        contract_traits::upgrade::<Self>(env, new_wasm_hash);
+        interfaces::upgrade::<Self>(env, new_wasm_hash);
     }
 }
 
@@ -54,7 +52,7 @@ impl UpgradableInterface for AxelarGasService {
 impl OwnableInterface for AxelarGasService {
     // boilerplate necessary for the contractimpl macro to include function in the generated client
     fn owner(env: &Env) -> Address {
-        contract_traits::owner(env)
+        interfaces::owner(env)
     }
 }
