+++ contracts/axelar-gateway/src/interface.rs
@@ -3,7 +3,7 @@ use crate::{
     types::{Message, Proof, WeightedSigners},
     AxelarGatewayMessagingInterface,
 };
-use axelar_soroban_std::contract_traits::UpgradableInterface;
+use axelar_soroban_std::interfaces::UpgradableInterface;
 use soroban_sdk::{contractclient, Address, BytesN, Env, Vec};
 
 #[contractclient(name = "AxelarGatewayClient")]
