--- packages/axelar-soroban-std/src/contract_traits/testdata/migrate_succeeds_if_owner_is_authenticated_and_called_after_upgrade.golden
@@ -1,3 +1,0 @@
-contract: Contract(CAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAD2KM)
-topics: (Symbol(upgraded))
-data: (String(0.1.0))
