+++ packages/axelar-soroban-std/src/interfaces/testdata/contract.wasm
