+++ packages/axelar-soroban-std/src/interfaces/testdata/migrate_succeeds_if_owner_is_authenticated_and_called_after_upgrade.golden
@@ -1,0 +1,1 @@
+Some(UpgradedEvent { version: String(0.1.0) })
